use module spacing to nest sequences in users
