+++ app/controllers/api/v1/asanas_controller.rb
@@ -1,5 +1,4 @@
 class Api::V1::AsanasController < ApplicationController
-skip_before_action :authorized, only: [:index,:show]
 	def index
 		render json: Asana.all
 	end
