+++ app/controllers/api/v1/categories_controller.rb
@@ -1,5 +1,4 @@
 class Api::V1::CategoriesController < ApplicationController
-skip_before_action :authorized, only: [:index,:show]
 	def index
 		render json: Category.all
 	end
