+++ config/routes.rb
@@ -3,8 +3,10 @@ Rails.application.routes.draw do
   namespace :api do
     namespace :v1 do
       resources :users, only: [:create,:index,:edit,:update,:destroy] do
-        resources :sequences, only: [:index,:show,:create,:update,:destroy]
-      end
+		scope module: 'users' do
+			resources :sequences, only: [:index,:show,:create,:update,:destroy]
+			end
+		end
       resources :categories, only: [:index,:show]
       resources :asanas, only: [:index,:show]
       post '/login', to: 'auth#create'
